requirements.txt
google-adk
python-dotenv

google-cloud-storage
google-cloud-firestore
fastapi
uvicorn
